requirements.txt
# Original dependencies
colorama==0.4.6
python_docx==0.8.11
python_pptx==0.6.21
Requests==2.31.0
setuptools==65.6.3
tqdm==4.64.1
sphinx_rtd_theme==1.2.2

# API Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=1.0.0
python-multipart>=0.0.6
